refactor: move table and cspf packages from internal/pkg to pkg
