+++ internal/pkg/gobgp/interface.go
@@ -18,7 +18,7 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/nttcom/pola/internal/pkg/table"
+	"github.com/nttcom/pola/pkg/table"
 	api "github.com/osrg/gobgp/v4/api"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
