+++ pkg/packet/pcep/message.go
@@ -11,7 +11,7 @@ import (
 	"fmt"
 	"net/netip"
 
-	"github.com/nttcom/pola/internal/pkg/table"
+	"github.com/nttcom/pola/pkg/table"
 )
 
 const CommonHeaderLength uint16 = 4
