+++ pkg/packet/pcep/object.go
@@ -12,7 +12,7 @@ import (
 	"math"
 	"net/netip"
 
-	"github.com/nttcom/pola/internal/pkg/table"
+	"github.com/nttcom/pola/pkg/table"
 )
 
 type PccType int
