+++ pkg/server/session.go
@@ -12,9 +12,9 @@ import (
 	"net/netip"
 	"time"
 
-	"github.com/nttcom/pola/internal/pkg/cspf"
-	"github.com/nttcom/pola/internal/pkg/table"
+	"github.com/nttcom/pola/pkg/cspf"
 	"github.com/nttcom/pola/pkg/packet/pcep"
+	"github.com/nttcom/pola/pkg/table"
 
 	"go.uber.org/zap"
 )
