+++ pkg/table/ted.go
